requirements/base.txt
# amqp==5.0.9
# asgiref==3.4.1
# attrs==21.4.0
# autopep8==1.6.0
# billiard==3.6.4.0
# botocore==1.23.29
# celery==5.2.3
# certifi==2021.10.8
# chardet==4.0.0
# charset-normalizer==2.0.10
# click==8.0.3
# click-didyoumean==0.3.0
# click-plugins==1.1.1
# click-repl==0.2.0
# cookies==2.2.1
# coreapi==2.3.3
# coreschema==0.0.4
# Django<4
# django-filter==21.1
# django-log-request-id==2.0.0
# djangorestframework==3.13.1
# djangorestframework-jwt==1.11.0
# docutils==0.18.1
# # drf-yasg==1.20.0
# graphene-django==2.15.0
# idna==3.3
# inflection==0.5.1
# itypes==1.2.0
# Jinja2==3.0.3
# jmespath==0.10.0
# kombu==5.2.3
# MarkupSafe==2.0.1
# model-mommy==2.0.0
# more-itertools==8.12.0
# openapi-codec==1.3.2
# packaging==21.3
# pbr==5.8.0
# pika==1.2.0
# pluggy==1.0.0
# promise==2.3
# prompt-toolkit==3.0.24
# psycopg2-binary==2.9.3
# py==1.11.0
# pycodestyle==2.8.0
# Pygments==2.11.2
# PyJWT==1.7.1
# pyparsing==3.0.6
# python-dateutil==2.8.2
# pytz==2021.3
# PyYAML==6.0
# redis==4.1.0
# requests==2.27.1
# responses==0.16.0
# ruamel.yaml==0.17.20
# ruamel.yaml.clib==0.2.6
# Rx==3.2.0
# s3transfer==0.5.0
# simplejson==3.17.6
# singledispatch==3.7.0
# six==1.16.0
# sqlparse==0.4.2
# text-unidecode==1.3
# toml==0.10.2
# uritemplate==4.1.1
# urllib3==1.26.7
# uWSGI==2.0.20
# vine==5.0.0
# wcwidth==0.2.5
amqp==5.0.9
asgiref==3.4.1
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
# botocore==1.23.29
celery==5.2.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cookies==2.2.1
coreapi==2.3.3
coreschema==0.0.4
Django<4
django-filter==21.1
django-log-request-id==2.0.0
djangorestframework>=3.13.1,!=3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt
# docutils==0.18.1
drf-yasg==1.20.0
graphene-django==2.15.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
# Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
# MarkupSafe==2.0.1
model-mommy==2.0.0
more-itertools==8.12.0
openapi-codec==1.3.2
packaging==21.3
pbr==5.8.0
pika==1.2.0
# pluggy==1.0.0
promise==2.3
prompt-toolkit==3.0.24
psycopg2-binary==2.9.3
py==1.11.0
# pycodestyle==2.8.0
Pygments==2.11.2
PyJWT==1.7.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.0
requests==2.27.1
responses==0.16.0
# ruamel.yaml==0.17.20
# ruamel.yaml.clib==0.2.6
Rx==3.2.0
# s3transfer==0.5.0
simplejson==3.17.6
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
uritemplate==4.1.1
# urllib3==1.26.7
# uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
# pyqt5<=5.15
# pyqtwebengine<5.13
pathlib
docutils<0.18,>=0.14
Jinja2<3.0,>=2.3
MarkupSafe<2.0
urllib3>=1.26.11
pluggy<1.0.0a1,>=0.12
pycodestyle>=2.8.0
# ,>=2.7.0
botocore<1.28.0,>=1.27.78
s3transfer<0.7.0,>=0.6.0